requirements.txt
tensorflow~=2.10.0
numpy~=1.23.3
flask~=2.1.3
keras~=2.10.0
h5py~=3.7.0
python-geohash~=0.8.5